refactor: remove usize casting

+++ src/parser/instruction/btype.rs
@@ -9,8 +9,8 @@ impl BType {
     pub fn execute(&self, cpu: &mut Processor, rs1: usize, rs2: usize, offset: i32) {
         match self {
             BType::Beq => {
-                let left = cpu.registers[rs1 as usize];
-                let right = cpu.registers[rs2 as usize];
+                let left = cpu.registers[rs1];
+                let right = cpu.registers[rs2];
                 if left == right {
                     let pc = cpu.pc as i32;
                     cpu.pc = (pc + offset) as usize;
